refactor: Change name productsReducer to productSlice
