+++ frontend/src/store.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { api } from './features/api/apiSlice';
-import { productsSlice } from './features/product/productReducers';
+import { productsSlice } from './features/product/productSlice';
 import authReducers from './features/auth/authReducers';
 import { cartSlice } from './features/cart/cartSlice';
 import { userInfoSlice } from './features/user/userReducers';
